refactor(cli): split benchmark.go into focused files to reduce complexity
Split the 3,769-line benchmark.go into 9 files organized by concern:
- benchmark.go: types, constants, command setup, main entry points
- benchmark_stats.go: statistical calculations (summary, percentile, mean)
- benchmark_report.go: report writing, GPU monitoring, nvidia-smi parsing
- benchmark_output.go: output formatting (table, JSON, markdown)
- benchmark_network.go: endpoint resolution, port-forwarding, k8s client
- benchmark_stress.go: stress test execution, prompt loading, HTTP requests
- benchmark_catalog.go: catalog benchmarking, model deployment/cleanup
- benchmark_sweep.go: parameter sweep logic (concurrency, tokens, context)
- benchmark_suite.go: test suite orchestration and phase execution

Extract helpers to bring all functions below gocyclo 15:
- initK8sClient: deduplicate k8s client initialization (4+ occurrences)
- runSweepIteration: deduplicate stress-vs-standard benchmark branching
- deployCatalogForPhase: consolidate deploy+wait+endpoint pattern
- printStressProgress, makeStopCondition, outputBenchmarkResults, etc.

Pure internal refactor with no behavioral or API changes.
All 103 existing tests pass unmodified.
