+++ pkg/cli/benchmark_output.go
@@ -1,0 +1,432 @@
+/*
+Copyright 2025.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package cli
+
+import (
+	"encoding/json"
+	"fmt"
+	"os"
+	"text/tabwriter"
+	"time"
+)
+
+func outputTable(summary BenchmarkSummary) {
+	fmt.Printf("📈 Benchmark Results\n")
+	fmt.Printf("═══════════════════════════════════════════════════════════════\n\n")
+
+	successRate := float64(summary.SuccessfulRuns) / float64(summary.Iterations) * 100
+	fmt.Printf("Runs: %d/%d successful (%.1f%%)\n\n",
+		summary.SuccessfulRuns, summary.Iterations, successRate)
+
+	if summary.SuccessfulRuns == 0 {
+		fmt.Printf("❌ No successful runs to report.\n")
+		return
+	}
+
+	w := tabwriter.NewWriter(os.Stdout, 0, 0, 2, ' ', 0)
+
+	_, _ = fmt.Fprintf(w, "THROUGHPUT\t\n")
+	_, _ = fmt.Fprintf(w, "──────────\t\n")
+	_, _ = fmt.Fprintf(w, "Generation:\t%.1f tok/s (mean)\t%.1f - %.1f tok/s (range)\n",
+		summary.GenerationToksPerSecMean,
+		summary.GenerationToksPerSecMin,
+		summary.GenerationToksPerSecMax)
+	if summary.PromptToksPerSecMean > 0 {
+		_, _ = fmt.Fprintf(w, "Prompt:\t%.1f tok/s (mean)\t\n", summary.PromptToksPerSecMean)
+	}
+	_ = w.Flush()
+
+	fmt.Println()
+
+	w = tabwriter.NewWriter(os.Stdout, 0, 0, 2, ' ', 0)
+	_, _ = fmt.Fprintf(w, "LATENCY\t\n")
+	_, _ = fmt.Fprintf(w, "───────\t\n")
+	_, _ = fmt.Fprintf(w, "P50:\t%.0f ms\t\n", summary.LatencyP50)
+	_, _ = fmt.Fprintf(w, "P95:\t%.0f ms\t\n", summary.LatencyP95)
+	_, _ = fmt.Fprintf(w, "P99:\t%.0f ms\t\n", summary.LatencyP99)
+	_, _ = fmt.Fprintf(w, "Min:\t%.0f ms\t\n", summary.LatencyMin)
+	_, _ = fmt.Fprintf(w, "Max:\t%.0f ms\t\n", summary.LatencyMax)
+	_, _ = fmt.Fprintf(w, "Mean:\t%.0f ms\t\n", summary.LatencyMean)
+	_ = w.Flush()
+
+	fmt.Printf("\n═══════════════════════════════════════════════════════════════\n")
+	fmt.Printf("Duration: %s\n", summary.Duration.Round(time.Second))
+	fmt.Printf("Prompt: %d tokens | Max generation: %d tokens\n",
+		summary.PromptTokens, summary.MaxTokens)
+}
+
+func outputJSON(summary BenchmarkSummary) error {
+	encoder := json.NewEncoder(os.Stdout)
+	encoder.SetIndent("", "  ")
+	return encoder.Encode(summary)
+}
+
+func outputMarkdown(summary BenchmarkSummary) {
+	fmt.Printf("# LLMKube Benchmark Results\n\n")
+	fmt.Printf("**Service:** %s  \n", summary.ServiceName)
+	fmt.Printf("**Namespace:** %s  \n", summary.Namespace)
+	fmt.Printf("**Date:** %s  \n\n", summary.Timestamp.Format("2006-01-02 15:04:05"))
+
+	successRate := float64(summary.SuccessfulRuns) / float64(summary.Iterations) * 100
+	fmt.Printf("## Summary\n\n")
+	fmt.Printf("| Metric | Value |\n")
+	fmt.Printf("|--------|-------|\n")
+	fmt.Printf("| Iterations | %d |\n", summary.Iterations)
+	fmt.Printf("| Success Rate | %.1f%% |\n", successRate)
+	fmt.Printf("| Duration | %s |\n\n", summary.Duration.Round(time.Second))
+
+	if summary.SuccessfulRuns == 0 {
+		fmt.Printf("No successful runs to report.\n")
+		return
+	}
+
+	fmt.Printf("## Throughput\n\n")
+	fmt.Printf("| Metric | Mean | Min | Max |\n")
+	fmt.Printf("|--------|------|-----|-----|\n")
+	fmt.Printf("| Generation (tok/s) | %.1f | %.1f | %.1f |\n",
+		summary.GenerationToksPerSecMean,
+		summary.GenerationToksPerSecMin,
+		summary.GenerationToksPerSecMax)
+	if summary.PromptToksPerSecMean > 0 {
+		fmt.Printf("| Prompt (tok/s) | %.1f | - | - |\n", summary.PromptToksPerSecMean)
+	}
+
+	fmt.Printf("\n## Latency\n\n")
+	fmt.Printf("| Percentile | Value (ms) |\n")
+	fmt.Printf("|------------|------------|\n")
+	fmt.Printf("| P50 | %.0f |\n", summary.LatencyP50)
+	fmt.Printf("| P95 | %.0f |\n", summary.LatencyP95)
+	fmt.Printf("| P99 | %.0f |\n", summary.LatencyP99)
+	fmt.Printf("| Min | %.0f |\n", summary.LatencyMin)
+	fmt.Printf("| Max | %.0f |\n", summary.LatencyMax)
+	fmt.Printf("| Mean | %.0f |\n", summary.LatencyMean)
+
+	fmt.Printf("\n---\n")
+	fmt.Printf("*Generated by LLMKube v%s*\n", Version)
+}
+
+func outputStressTable(summary StressTestSummary) {
+	fmt.Printf("📈 Stress Test Results\n")
+	fmt.Printf("═══════════════════════════════════════════════════════════════\n\n")
+
+	fmt.Printf("OVERVIEW\n")
+	fmt.Printf("────────\n")
+	fmt.Printf("Total Requests:  %d\n", summary.TotalRequests)
+	fmt.Printf("Success Rate:    %.1f%% (%d/%d)\n",
+		100-summary.ErrorRate, summary.SuccessfulRuns, summary.TotalRequests)
+	fmt.Printf("Duration:        %s\n", summary.Duration.Round(time.Second))
+	fmt.Printf("Concurrency:     %d\n", summary.Concurrency)
+	fmt.Printf("Requests/sec:    %.2f\n\n", summary.RequestsPerSec)
+
+	if summary.SuccessfulRuns == 0 {
+		fmt.Printf("❌ No successful runs to report.\n")
+		return
+	}
+
+	w := tabwriter.NewWriter(os.Stdout, 0, 0, 2, ' ', 0)
+	_, _ = fmt.Fprintf(w, "THROUGHPUT\t\n")
+	_, _ = fmt.Fprintf(w, "──────────\t\n")
+	_, _ = fmt.Fprintf(w, "Generation:\t%.1f tok/s (mean)\t%.1f - %.1f (range)\n",
+		summary.GenerationToksPerSecMean,
+		summary.GenerationToksPerSecMin,
+		summary.GenerationToksPerSecMax)
+	_, _ = fmt.Fprintf(w, "Peak:\t%.1f tok/s\t\n", summary.PeakToksPerSec)
+	if summary.PromptToksPerSecMean > 0 {
+		_, _ = fmt.Fprintf(w, "Prompt:\t%.1f tok/s (mean)\t\n", summary.PromptToksPerSecMean)
+	}
+	_ = w.Flush()
+
+	fmt.Println()
+
+	w = tabwriter.NewWriter(os.Stdout, 0, 0, 2, ' ', 0)
+	_, _ = fmt.Fprintf(w, "LATENCY\t\n")
+	_, _ = fmt.Fprintf(w, "───────\t\n")
+	_, _ = fmt.Fprintf(w, "P50:\t%.0f ms\t\n", summary.LatencyP50)
+	_, _ = fmt.Fprintf(w, "P95:\t%.0f ms\t\n", summary.LatencyP95)
+	_, _ = fmt.Fprintf(w, "P99:\t%.0f ms\t\n", summary.LatencyP99)
+	_, _ = fmt.Fprintf(w, "Min:\t%.0f ms\t\n", summary.LatencyMin)
+	_, _ = fmt.Fprintf(w, "Max:\t%.0f ms\t\n", summary.LatencyMax)
+	_, _ = fmt.Fprintf(w, "Mean:\t%.0f ms\t\n", summary.LatencyMean)
+	_ = w.Flush()
+
+	fmt.Printf("\n═══════════════════════════════════════════════════════════════\n")
+	fmt.Printf("Max tokens per request: %d\n", summary.MaxTokens)
+}
+
+func outputStressJSON(summary StressTestSummary) error {
+	encoder := json.NewEncoder(os.Stdout)
+	encoder.SetIndent("", "  ")
+	return encoder.Encode(summary)
+}
+
+func outputStressMarkdown(summary StressTestSummary) {
+	fmt.Printf("# LLMKube Stress Test Results\n\n")
+	fmt.Printf("**Service:** %s  \n", summary.ServiceName)
+	fmt.Printf("**Namespace:** %s  \n", summary.Namespace)
+	fmt.Printf("**Date:** %s  \n\n", summary.Timestamp.Format("2006-01-02 15:04:05"))
+
+	fmt.Printf("## Overview\n\n")
+	fmt.Printf("| Metric | Value |\n")
+	fmt.Printf("|--------|-------|\n")
+	fmt.Printf("| Total Requests | %d |\n", summary.TotalRequests)
+	fmt.Printf("| Success Rate | %.1f%% |\n", 100-summary.ErrorRate)
+	fmt.Printf("| Duration | %s |\n", summary.Duration.Round(time.Second))
+	fmt.Printf("| Concurrency | %d |\n", summary.Concurrency)
+	fmt.Printf("| Requests/sec | %.2f |\n\n", summary.RequestsPerSec)
+
+	if summary.SuccessfulRuns == 0 {
+		fmt.Printf("No successful runs to report.\n")
+		return
+	}
+
+	fmt.Printf("## Throughput\n\n")
+	fmt.Printf("| Metric | Mean | Min | Max | Peak |\n")
+	fmt.Printf("|--------|------|-----|-----|------|\n")
+	fmt.Printf("| Generation (tok/s) | %.1f | %.1f | %.1f | %.1f |\n",
+		summary.GenerationToksPerSecMean,
+		summary.GenerationToksPerSecMin,
+		summary.GenerationToksPerSecMax,
+		summary.PeakToksPerSec)
+	if summary.PromptToksPerSecMean > 0 {
+		fmt.Printf("| Prompt (tok/s) | %.1f | - | - | - |\n", summary.PromptToksPerSecMean)
+	}
+
+	fmt.Printf("\n## Latency\n\n")
+	fmt.Printf("| Percentile | Value (ms) |\n")
+	fmt.Printf("|------------|------------|\n")
+	fmt.Printf("| P50 | %.0f |\n", summary.LatencyP50)
+	fmt.Printf("| P95 | %.0f |\n", summary.LatencyP95)
+	fmt.Printf("| P99 | %.0f |\n", summary.LatencyP99)
+	fmt.Printf("| Min | %.0f |\n", summary.LatencyMin)
+	fmt.Printf("| Max | %.0f |\n", summary.LatencyMax)
+	fmt.Printf("| Mean | %.0f |\n", summary.LatencyMean)
+
+	fmt.Printf("\n---\n")
+	fmt.Printf("*Generated by LLMKube v%s*\n", Version)
+}
+
+func printComparisonConfigLine(report ComparisonReport) {
+	if report.IsStressTest {
+		accel := acceleratorCPU
+		if report.GPUEnabled {
+			accel = fmt.Sprintf("%s | GPU Count: %d", report.Accelerator, report.GPUCount)
+		}
+		if report.TargetDuration > 0 {
+			fmt.Printf("Accelerator: %s | Concurrency: %d | Duration: %s | Max Tokens: %d\n\n",
+				accel, report.Concurrency, report.TargetDuration, report.MaxTokens)
+		} else {
+			fmt.Printf("Accelerator: %s | Concurrency: %d | Iterations: %d | Max Tokens: %d\n\n",
+				accel, report.Concurrency, report.Iterations, report.MaxTokens)
+		}
+		return
+	}
+	if report.GPUEnabled {
+		fmt.Printf("Accelerator: %s | GPU Count: %d | Iterations: %d | Max Tokens: %d\n\n",
+			report.Accelerator, report.GPUCount, report.Iterations, report.MaxTokens)
+	} else {
+		fmt.Printf("Accelerator: %s | Iterations: %d | Max Tokens: %d\n\n",
+			acceleratorCPU,
+			report.Iterations, report.MaxTokens)
+	}
+}
+
+func writeStressModelRow(w *tabwriter.Writer, m ModelBenchmark) {
+	status := statusIconSuccess
+	if m.Status != statusSuccess {
+		status = statusIconFailed
+	}
+	requests, rps, tps, p50, p99, errRate := "-", "-", "-", "-", "-", "-"
+	if m.Status == statusSuccess {
+		requests = fmt.Sprintf("%d", m.TotalRequests)
+		rps = fmt.Sprintf("%.1f", m.RequestsPerSec)
+		tps = fmt.Sprintf("%.1f", m.GenerationToksPerSec)
+		p50 = fmt.Sprintf("%.0f", m.LatencyP50Ms)
+		p99 = fmt.Sprintf("%.0f", m.LatencyP99Ms)
+		errRate = fmt.Sprintf("%.1f", m.ErrorRate)
+	}
+	_, _ = fmt.Fprintf(w, "%s\t%s\t%s\t%s\t%s\t%s\t%s\t%s\t%s\n",
+		m.ModelID, m.ModelSize, requests, rps, tps, p50, p99, errRate, status)
+}
+
+func writeStandardModelRow(w *tabwriter.Writer, m ModelBenchmark) {
+	status := statusIconSuccess
+	if m.Status != statusSuccess {
+		status = statusIconFailed
+	}
+	genToks, p50, p99 := "-", "-", "-"
+	if m.Status == statusSuccess {
+		genToks = fmt.Sprintf("%.1f", m.GenerationToksPerSec)
+		p50 = fmt.Sprintf("%.0f", m.LatencyP50Ms)
+		p99 = fmt.Sprintf("%.0f", m.LatencyP99Ms)
+	}
+	_, _ = fmt.Fprintf(w, "%s\t%s\t%s\t%s\t%s\t%s\t%s\n",
+		m.ModelID, m.ModelSize, genToks, p50, p99, m.VRAMEstimate, status)
+}
+
+func outputComparisonTable(report ComparisonReport) error {
+	if report.IsStressTest {
+		fmt.Printf("📊 Stress Test Comparison Results\n")
+	} else {
+		fmt.Printf("📊 Benchmark Comparison Results\n")
+	}
+	fmt.Printf("═══════════════════════════════════════════════════════════════════════════════\n\n")
+
+	successes := 0
+	for _, m := range report.Models {
+		if m.Status == statusSuccess {
+			successes++
+		}
+	}
+	fmt.Printf("Models: %d/%d benchmarked successfully\n", successes, len(report.Models))
+
+	printComparisonConfigLine(report)
+
+	w := tabwriter.NewWriter(os.Stdout, 0, 0, 2, ' ', 0)
+	if report.IsStressTest {
+		_, _ = fmt.Fprintf(w, "MODEL\tSIZE\tREQUESTS\tREQ/S\tTOK/S\tP50 (ms)\tP99 (ms)\tERROR%%\tSTATUS\n")
+		_, _ = fmt.Fprintf(w, "─────\t────\t────────\t─────\t─────\t────────\t────────\t──────\t──────\n")
+		for _, m := range report.Models {
+			writeStressModelRow(w, m)
+		}
+	} else {
+		_, _ = fmt.Fprintf(w, "MODEL\tSIZE\tGEN TOK/S\tP50 (ms)\tP99 (ms)\tVRAM\tSTATUS\n")
+		_, _ = fmt.Fprintf(w, "─────\t────\t─────────\t────────\t────────\t────\t──────\n")
+		for _, m := range report.Models {
+			writeStandardModelRow(w, m)
+		}
+	}
+	_ = w.Flush()
+
+	hasErrors := false
+	for _, m := range report.Models {
+		if m.Error != "" {
+			if !hasErrors {
+				fmt.Printf("\n⚠️  Errors:\n")
+				hasErrors = true
+			}
+			fmt.Printf("   %s: %s\n", m.ModelID, m.Error)
+		}
+	}
+
+	fmt.Printf("\n═══════════════════════════════════════════════════════════════════════════════\n")
+	fmt.Printf("Total Duration: %s\n", report.Duration.Round(time.Second))
+
+	return nil
+}
+
+func outputComparisonJSON(report ComparisonReport) error {
+	encoder := json.NewEncoder(os.Stdout)
+	encoder.SetIndent("", "  ")
+	return encoder.Encode(report)
+}
+
+func outputComparisonMarkdown(report ComparisonReport) error {
+	fmt.Printf("# LLMKube Benchmark Comparison\n\n")
+	fmt.Printf("**Date:** %s  \n", report.Timestamp.Format("2006-01-02 15:04:05"))
+	fmt.Printf("**GPU Enabled:** %v  \n", report.GPUEnabled)
+	fmt.Printf("**Iterations:** %d per model  \n", report.Iterations)
+	fmt.Printf("**Max Tokens:** %d  \n\n", report.MaxTokens)
+
+	fmt.Printf("## Results\n\n")
+	fmt.Printf("| Model | Size | Gen tok/s | P50 (ms) | P99 (ms) | VRAM | Status |\n")
+	fmt.Printf("|-------|------|-----------|----------|----------|------|--------|\n")
+
+	for _, m := range report.Models {
+		status := "✅ Success"
+		if m.Status != statusSuccess {
+			status = "❌ Failed"
+		}
+
+		genToks := "-"
+		p50 := "-"
+		p99 := "-"
+		if m.Status == statusSuccess {
+			genToks = fmt.Sprintf("%.1f", m.GenerationToksPerSec)
+			p50 = fmt.Sprintf("%.0f", m.LatencyP50Ms)
+			p99 = fmt.Sprintf("%.0f", m.LatencyP99Ms)
+		}
+
+		fmt.Printf("| %s | %s | %s | %s | %s | %s | %s |\n",
+			m.ModelID,
+			m.ModelSize,
+			genToks,
+			p50,
+			p99,
+			m.VRAMEstimate,
+			status,
+		)
+	}
+
+	hasErrors := false
+	for _, m := range report.Models {
+		if m.Error != "" {
+			if !hasErrors {
+				fmt.Printf("\n## Errors\n\n")
+				hasErrors = true
+			}
+			fmt.Printf("- **%s**: %s\n", m.ModelID, m.Error)
+		}
+	}
+
+	fmt.Printf("\n---\n")
+	fmt.Printf("*Total Duration: %s*  \n", report.Duration.Round(time.Second))
+	fmt.Printf("*Generated by LLMKube v%s*\n", Version)
+
+	return nil
+}
+
+func outputSweepTable(report SweepReport) {
+	fmt.Printf("\n📊 %s Sweep Results\n", report.SweepType)
+	fmt.Printf("═══════════════════════════════════════════════════════════════════════════════\n\n")
+
+	w := tabwriter.NewWriter(os.Stdout, 0, 0, 2, ' ', 0)
+	_, _ = fmt.Fprintf(w, "VALUE\tGEN TOK/S\tP50 (ms)\tP99 (ms)\tREQUESTS\tRPS\tERROR%%\tSTATUS\n")
+	_, _ = fmt.Fprintf(w, "─────\t─────────\t────────\t────────\t────────\t───\t──────\t──────\n")
+
+	for _, r := range report.Results {
+		status := statusIconSuccess
+		genToks := "-"
+		p50 := "-"
+		p99 := "-"
+		requests := "-"
+		rps := "-"
+		errRate := "-"
+
+		if r.Error != "" {
+			status = statusIconFailed
+		} else if r.Stress != nil {
+			genToks = fmt.Sprintf("%.1f", r.Stress.GenerationToksPerSecMean)
+			p50 = fmt.Sprintf("%.0f", r.Stress.LatencyP50)
+			p99 = fmt.Sprintf("%.0f", r.Stress.LatencyP99)
+			requests = fmt.Sprintf("%d", r.Stress.TotalRequests)
+			rps = fmt.Sprintf("%.1f", r.Stress.RequestsPerSec)
+			errRate = fmt.Sprintf("%.1f", r.Stress.ErrorRate)
+		} else if r.Summary != nil {
+			genToks = fmt.Sprintf("%.1f", r.Summary.GenerationToksPerSecMean)
+			p50 = fmt.Sprintf("%.0f", r.Summary.LatencyP50)
+			p99 = fmt.Sprintf("%.0f", r.Summary.LatencyP99)
+			requests = fmt.Sprintf("%d", r.Summary.Iterations)
+		}
+
+		_, _ = fmt.Fprintf(w, "%s\t%s\t%s\t%s\t%s\t%s\t%s\t%s\n",
+			r.Value, genToks, p50, p99, requests, rps, errRate, status)
+	}
+	_ = w.Flush()
+
+	fmt.Printf("\n═══════════════════════════════════════════════════════════════════════════════\n")
+	fmt.Printf("Total Duration: %s\n", report.Duration.Round(time.Second))
+}
